pyproject.toml
[project]
name = "offers-marketplace-backend"
version = "0.1.0"
description = "Serverless FastAPI backend with AWS CDK infrastructure"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.138",
    "fastapi>=0.111.0",
    "mangum>=0.17.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.5.0",
    "uvicorn[standard]>=0.30.0",
]
infra = [
    "aws-cdk-lib>=2.149.0",
    "constructs>=10.3.0",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py312"
line-length = 99

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"
